Update integration with library
