+++ CSL Scrollable Toolbar/CSL Scrollable Toolbar.csproj
@@ -55,7 +55,6 @@
     <Compile Include="Configuration.cs" />
     <Compile Include="Defs\GameObjectDefs.cs" />
     <Compile Include="Detour\CustomMouseHandler.cs" />
-    <Compile Include="LoadingExtension.cs" />
     <Compile Include="Mod.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UI\Toolbar.cs" />
